ingestion_pipeline/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
automat==25.4.16
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-option-group==0.5.7
constantly==23.10.4
cryptography==46.0.5
cssselect==1.4.0
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
defusedxml==0.7.1
distro==1.9.0
docstring-parser==0.17.0
filelock==3.25.2
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-aiplatform==1.142.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.68.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.7.2
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
kfp==2.16.0
kfp-pipeline-spec==2.16.0
kfp-server-api==2.16.0
kubernetes==30.1.0
langchain==1.2.13
langchain-core==1.2.22
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.2
markupsafe==3.0.3
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
oauthlib==3.3.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
parsel==1.11.0
pinecone-text==0.11.0
protego==0.6.0
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydispatcher==2.0.7
pygments==2.19.2
pyopenssl==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyyaml==6.0.3
queuelib==1.9.0
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
scrapy==2.14.2
sentence-transformers==5.3.0
service-identity==24.2.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tldextract==5.3.1
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.3.0
triton==3.6.0
twisted==25.5.0
typer==0.24.1
types-requests==2.32.4.20260324
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
w3lib==2.4.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
zope-interface==8.2
zstandard==0.25.0
